requirements.txt
Django==4.0.2
wagtail==2.16.1
djangorestframework==3.13.1
